+++ src/components/Buttons/index.ts
@@ -1,8 +1,10 @@
 export { default as ActionButton } from './ActionButton';
 export { default as BackButton } from './back-button';
+export { default as BigButton } from './big-button';
+export { default as BrandedButton } from './branded-button';
 export { default as CheckBoxButton } from './check-box-button';
-export { default as IconButton } from './IconButton';
+export { default as IconButton } from './icon-button';
 export { default as Link } from './link';
 export { default as RoundIconButton } from './round-icon-button';
-export { default as ShareButton } from './ShareButton';
-export { default as ThemeButton } from './ThemeButton';
+export { default as ShareButton } from './share-button';
+export { default as ThemeButton } from './theme-button';
