+++ src/components/Cards/simple-share.tsx
@@ -4,8 +4,7 @@ import { Image, StyleSheet, View } from 'react-native';
 import { social } from '@assets';
 import i18n from '@covid/locale/i18n';
 import { track } from '@covid/core/Analytics';
-
-import { BrandedButton } from '../BrandedButton';
+import { BrandedButton } from '@covid/components';
 
 import { share } from './BaseShareApp';
