+++ src/features/covid-tests/NHSTestDetailScreen.tsx
@@ -8,7 +8,7 @@ import * as Yup from 'yup';
 
 import ProgressStatus from '@covid/components/ProgressStatus';
 import Screen, { Header, ProgressBlock } from '@covid/components/Screen';
-import { BrandedButton, ErrorText, HeaderText } from '@covid/components/Text';
+import { ErrorText, HeaderText } from '@covid/components/Text';
 import { ValidationError } from '@covid/components/ValidationError';
 import { ICovidTestService } from '@covid/core/user/CovidTestService';
 import { CovidTest, CovidTestType } from '@covid/core/user/dto/CovidTestContracts';
@@ -19,21 +19,16 @@ import { ScreenParamList } from '@covid/features/ScreenParamList';
 import { Services } from '@covid/provider/services.types';
 import { lazyInject } from '@covid/provider/services';
 import { ClearButton } from '@covid/components/Buttons/ClearButton';
+import { INHSTestDateData, NHSTestDateQuestion } from '@covid/features/covid-tests/fields/NHSTestDateQuestion';
 import {
-  INHSTestDateData,
-  INHSTestDateQuestion,
-  NHSTestDateQuestion,
-} from '@covid/features/covid-tests/fields/NHSTestDateQuestion';
-import {
-  CovidTestTimeQuestion,
-  ICovidTestTimeData,
-  ICovidTestResultData,
-  ICovidTestResultQuestion,
-  INHSTestMechanismData,
-  INHSTestMechanismQuestion,
-  NHSTestMechanismQuestion,
   CovidTestResultQuestion,
+  CovidTestTimeQuestion,
+  ICovidTestResultData,
+  ICovidTestTimeData,
+  INHSTestMechanismData,
+  NHSTestMechanismQuestion,
 } from '@covid/features/covid-tests/fields/';
+import { BrandedButton } from '@covid/components';
 
 interface INHSTestData extends INHSTestDateData, INHSTestMechanismData, ICovidTestResultData, ICovidTestTimeData {}
