+++ src/features/multi-profile/ArchiveReasonScreen.tsx
@@ -4,14 +4,13 @@ import { StackNavigationProp } from '@react-navigation/stack';
 
 import { HeaderText, RegularText, SecondaryText } from '@covid/components/Text';
 import Screen, { FieldWrapper, Header } from '@covid/components/Screen';
-import { BigButton } from '@covid/components/BigButton';
 import i18n from '@covid/locale/i18n';
 import { useInjection } from '@covid/provider/services.hooks';
 import { Services } from '@covid/provider/services.types';
 import { IPatientService } from '@covid/core/patient/PatientService';
 import appCoordinator from '@covid/features/AppCoordinator';
-
-import { ScreenParamList } from '../ScreenParamList';
+import { ScreenParamList } from '@covid/features';
+import { BigButton } from '@covid/components';
 
 type RenderProps = {
   navigation: StackNavigationProp<ScreenParamList, 'ArchiveReason'>;
